this seems to compile

+++ src/Demes.hpp
@@ -73,11 +73,11 @@ private:
   std::vector<Eigen::MatrixXd> pulses_;
 
   // following (outer) vectors store one object per epoch per population:
-  std::vector<std::vector<long double>> mutRates_;
-  std::vector<std::vector<long double>> recRates_;
-  std::vector<std::vector<long double>> selCoeffs_;
+  std::vector<std::vector<double>> mutRates_;
+  std::vector<std::vector<double>> recRates_;
+  std::vector<std::vector<double>> selCoeffs_;
 
-  long double leftFactor_; // see Mutation.hpp
+  double leftFactor_; // see Mutation.hpp
 
 public:
   Demes(const std::string& file):
@@ -153,17 +153,17 @@ public:
     return pops_[epoch].size();
   }
 
-  const std::vector<long double>& getMus(size_t epoch) const
+  const std::vector<double>& getMus(size_t epoch) const
   {
     return mutRates_[epoch];
   }
 
-  const std::vector<long double>& getRecs(size_t epoch) const
+  const std::vector<double>& getRecs(size_t epoch) const
   {
     return recRates_[epoch];
   }
 
-  const std::vector<long double>& getSelCoeffs(size_t epoch) const
+  const std::vector<double>& getSelCoeffs(size_t epoch) const
   {
     return selCoeffs_[epoch];
   }
@@ -188,18 +188,18 @@ public:
     return pulses_[epoch];
   }
 
-  long double getLeftFactor()
+  double getLeftFactor()
   {
     return leftFactor_;
   }
 
-  void setMus(size_t epoch, long double mu)
+  void setMus(size_t epoch, double mu)
   {
     for(size_t i = 0; i < pops_[i].size(); ++i)
       mutRates_[epoch][i] = mu;
   }
 
-  void setRecs(size_t epoch, long double rec)
+  void setRecs(size_t epoch, double rec)
   {
     for(size_t i = 0; i < pops_[i].size(); ++i)
       recRates_[epoch][i] = rec;
