+++ src/Selection.hpp
@@ -23,13 +23,12 @@ public:
     setUpMatrices_(sslib);
   }
 
-  Selection(const std::vector<long double>& vals, std::shared_ptr<bpp::IntervalConstraint> ic, const SumStatsLibrary& sslib):
+  Selection(const std::vector<double>& vals, std::shared_ptr<bpp::IntervalConstraint> ic, const SumStatsLibrary& sslib):
   AbstractOperator(sslib.getPopIndices())
   {
     // for each population modeled in the epoch *this operator belongs to, add s parameter
     for(size_t i = 0; i < popIndices_.size(); ++i)
-      addParameter_(
-          new bpp::Parameter("s_" + bpp::TextTools::toString(popIndices_[i]), vals[i], ic));
+      addParameter_(new bpp::Parameter("s_" + bpp::TextTools::toString(popIndices_[i]), vals[i], ic));
 
     prevParams_.addParameters(getParameters()); // inits list of "previous" parameters
     setUpMatrices_(sslib);
